Extract DPM command-center API module
